refactor(sec): extract duplicated SGML tag-value scan into SecSgmlEnvelope helper
